feat: initialize @withstudiocms/config-utils package with core utilities for configuration management
- Add .gitignore to exclude generated files and dependencies
- Include MIT License for open-source distribution
- Create README.md with package description and licensing information
- Set up package.json with metadata, scripts, and dependencies
- Implement core utilities in src/loader.ts, src/watcher.ts, and src/zod-utils.ts for loading, watching, and validating configuration files
- Add utility functions for dynamic result handling in src/utils/
- Configure TypeScript with tsconfig.json and tsconfig.tspc.json for build and type checking
