+++ packages/@withstudiocms/config-utils/README.md
@@ -1,0 +1,7 @@
+# @withstudiocms/config-utils
+
+Collection of utilities for working with config files and Astro integrations.
+
+## License
+
+[MIT Licensed](./LICENSE)
