+++ packages/@withstudiocms/config-utils/tsconfig.tspc.json
@@ -1,0 +1,9 @@
+{
+  "extends": "../../../tsconfig.tspc.json",
+  "include": ["src"],
+  "exclude": ["dist"],
+  "compilerOptions": {
+    "outDir": "./dist",
+    "rootDir": "./src"
+  }
+}
