+++ packages/studiocms/package.json
@@ -170,18 +170,17 @@
 		"@oslojs/encoding": "^1.1.0",
 		"@studiocms/ui": "catalog:studiocms",
 		"@withstudiocms/cli-kit": "catalog:studiocms",
+		"@withstudiocms/config-utils": "workspace:*",
 		"arctic": "^3.7.0",
 		"astro-integration-kit": "catalog:",
 		"boxen": "^8.0.1",
 		"chalk": "^5.4.1",
-		"deepmerge-ts": "^7.1.5",
 		"diff": "^8.0.2",
 		"diff2html": "^3.4.52",
 		"dompurify": "^3.2.6",
 		"dotenv": "^17.2.1",
 		"drizzle-orm": "^0.42.0",
 		"effect": "catalog:effect",
-		"esbuild": "^0.25.6",
 		"fuse.js": "^7.1.0",
 		"jose": "^6.0.12",
 		"markdown-it": "^14.1.0",
